refactor(components): update import path for `cn` utility function
Change import path of `cn` from "@/lib/cn" to "@/lib/utils" across
multiple components to reflect the new location of the utility
function. This change centralizes utility functions, improving
maintainability and reducing redundancy.

feat(utils): add utility function for class name merging

Introduce a new utility function `cn` in `utils.ts` that combines
`clsx` and `tailwind-merge` for efficient class name merging. This
enhancement provides a more robust solution for handling conditional
class names and tailwind class merging.
