+++ apps/web/src/components/ui/form.tsx
@@ -14,7 +14,7 @@ import {
 } from "react-hook-form";
 
 import { Label } from "@/components/ui/label";
-import { cn } from "@/lib/cn";
+import { cn } from "@/lib/utils";
 
 const Slot = SlotPrimitive.Slot;
