+++ apps/web/src/components/ui/table.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { cn } from "@/lib/cn";
+import { cn } from "@/lib/utils";
 
 function Table({ className, ...props }: React.ComponentProps<"table">) {
   return (
