+++ apps/web/src/components/ui/tabs.tsx
@@ -3,7 +3,7 @@
 import { Tabs as TabsPrimitive } from "radix-ui";
 import * as React from "react";
 
-import { cn } from "@/lib/cn";
+import { cn } from "@/lib/utils";
 
 function Tabs({
   className,
