+++ apps/web/src/components/ui/typography.tsx
@@ -2,7 +2,7 @@ import { LinkIcon } from "lucide-react";
 import { Slot as SlotPrimitive } from "radix-ui";
 import React from "react";
 
-import { cn } from "@/lib/cn";
+import { cn } from "@/lib/utils";
 
 const Slot = SlotPrimitive.Slot;
