+++ apps/web/src/components/wheel-picker.tsx
@@ -2,7 +2,7 @@ import "@ncdai/react-wheel-picker/style.css";
 
 import * as WheelPickerPrimitive from "@ncdai/react-wheel-picker";
 
-import { cn } from "@/lib/cn";
+import { cn } from "@/lib/utils";
 
 type WheelPickerOption = WheelPickerPrimitive.WheelPickerOption;
 type WheelPickerClassNames = WheelPickerPrimitive.WheelPickerClassNames;
@@ -14,10 +14,10 @@ function WheelPickerWrapper({
   return (
     <WheelPickerPrimitive.WheelPickerWrapper
       className={cn(
-        "w-56 rounded-lg bg-white px-1 shadow-sm ring ring-black/5 dark:bg-zinc-900 dark:ring-white/15",
+        "w-56 rounded-lg border border-zinc-200 bg-white px-1 shadow-xs dark:border-zinc-700/80 dark:bg-zinc-900",
         "*:data-rwp:first:*:data-rwp-highlight-wrapper:rounded-s-md",
         "*:data-rwp:last:*:data-rwp-highlight-wrapper:rounded-e-md",
-        className,
+        className
       )}
       {...props}
     />
