+++ apps/web/src/lib/utils.ts
@@ -1,4 +1,4 @@
-import type { ClassValue} from "clsx";
+import type { ClassValue } from "clsx";
 import { clsx } from "clsx";
 import { twMerge } from "tailwind-merge";
